Tweaking Runner to plot resulting values on canvas
git-svn-id: svn://localhost/trunk@4073 a386ff7f-d321-c144-bb17-e251d28fee14

+++ Utils/Runner/Unit_Runner.pas
@@ -11,7 +11,7 @@ type
   TKMRunnerClass = class of TKMRunnerCommon;
 
   TKMRunResult = record
-    Value: Integer;
+    Value: Single;
   end;
 
   TKMRunResults = array of TKMRunResult;
